requirements.txt
beautifulsoup4==4.9.3
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
cymem==2.0.3
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.1/en_core_web_md-2.3.1.tar.gz
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
idna==2.10
jieba3k==0.35.1
joblib==0.17.0
lxml==4.6.1
murmurhash==1.0.2
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.2
Pillow==8.0.1
plac==1.1.3
preshed==3.0.2
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.10.23
requests==2.24.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
spacy==2.3.1
srsly==1.0.2
thinc==7.4.1
tinysegmenter==0.3
tldextract==3.0.2
tqdm==4.51.0
urllib3==1.25.11
wasabi==0.8.0
